pyproject.toml
[tool.poetry]
name = "procon-grassmaker"
version = "0.0.1"
description = ""
authors = ["Masaki Kobayashi <bayashi.cl@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
bs4 = "^0.0.1"
GitPython = "^3.1.24"
python-dateutil = "^2.8.2"
toml = "^0.10.2"
colorama = "^0.4.4"
colorlog = "^6.6.0"
dacite = "^1.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.930"
isort = "^5.10.1"
black = "^21.12b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
procon-grassmaker = "procon_grassmaker:main"
